player.js : added drag notes in timeline functionality
Added note dragging functionality.

TODO: Dragging earlier than parent startTime dependency, update dependency to the dependencies parent : allowing us to move notes up until the BaseNote.
+
detect when parent dependency is a measure bar and if we move further than the measure length search for a child measure bar that we can link to and update dependencies if it exists.
+
improve drag to represent distance in tapspace and not fixed drag distance.